requirements.txt
Flask==2.3.3
b2sdk>=2.0.0
python-dotenv==1.0.0
flask-cors>=6.0.0
